Removed a bug and some small changes
Removed a bug in "is_authenticated" function that caused to
return Ture in some cases with even if only the email or the
password are found in the database. I also passed 'cache_id'
varibale to the templates to prevent caching some files. Finally,
I added a function to renew the database session after each serve.